IGNITE-20272 The fix after review from Mirza.
